pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "Complete MLOps Pipeline for ML Practationers"
authors = [
  { name = "Ezzaldin" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "dagshub~=0.3.30",
    "duckdb>=1.2.2",
    "dvc~=3.33.0",
    "dvc-s3~=3.2.0",
    "hydra-core~=1.3.0",
    "hyperopt~=0.2.5",
    "litserve~=0.2.8",
    "loguru~=0.7.2",
    "mlflow~=2.20.0",
    "numpy~=1.26.4",
    "pandas~=2.0.1",
    "pip",
    "prefect>=3.4.0",
    "pyarrow~=19.0.0",
    "python-dotenv",
    "ruff",
    "scikit-learn~=1.6.0",
    "skore~=0.8.2",
    "sktime>=0.37.0",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "kagglehub>=0.3.11",
]
linting = [
    "black>=25.1.0",
    "isort>=6.0.1",
]
optional = [
    "hydra-colorlog>=1.2.0",
    "wandb~=0.19.6",
]

